# docs/swagger.yaml
basePath: /v1
definitions:
  datatransfers.Answer:
    properties:
      questionID:
        type: integer
      value:
        type: string
    type: object
  models.Question:
    properties:
      answer:
        type: string
      created_at:
        type: integer
      deleted_at:
        type: integer
      id:
        type: integer
      question:
        type: string
      updated_at:
        type: integer
    type: object
info:
  contact:
    email: reziapriliansyah@gmail.com
    name: Rezi Apriliansyah
  description: API documentation for quiz master
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Api Documentation for quiz master
  version: 1.0.0
paths:
  /questions:
    get:
      description: Get the list of questions
      operationId: list-questions
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: number of questions in single page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Question'
            type: array
      summary: List questions
      tags:
      - questions
    post:
      description: Create a new questions
      operationId: create-questions
      parameters:
      - description: Create questions
        in: body
        name: questions
        required: true
        schema:
          $ref: '#/definitions/models.Question'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Question'
      summary: Create a new questions
      tags:
      - questions
  /questions/{id}:
    delete:
      description: Delete an questions by id
      operationId: delete-questions
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
          schema:
            type: string
      summary: Delete an questions
      tags:
      - questions
    get:
      description: Get a question by id
      operationId: get-question-id
      parameters:
      - description: id Question
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Question'
      summary: Get a question by id
      tags:
      - questions
    put:
      description: Update an questions by id
      operationId: update-questions
      parameters:
      - description: Update questions
        in: body
        name: questions
        required: true
        schema:
          $ref: '#/definitions/models.Question'
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Question'
      summary: Update an questions
      tags:
      - questions
  /questions/answer/{id}:
    post:
      description: Answer a new questions
      operationId: answer-questions
      parameters:
      - description: Answer questions
        in: body
        name: questions
        required: true
        schema:
          $ref: '#/definitions/datatransfers.Answer'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Answer a new questions
      tags:
      - questions
swagger: "2.0"